Fix metalinter complaints
